+++ page-components/Post/Poster.jsx
@@ -1,9 +1,10 @@
-import { Avatar } from '@/components/Avatar';
 import { LoadingDots } from '@/components/LoadingDots';
 import { Text, TextLink } from '@/components/Text';
 import Link from 'next/link';
 import styles from './Poster.module.css';
 import { PosterInner } from '@/components/PosterInner';
+import { UserOutlined } from '@ant-design/icons';
+import { Avatar } from 'antd';
 
 const Poster = ({ data, error, isCreate }) => {
   const loading = !data && !error;
@@ -11,7 +12,13 @@ const Poster = ({ data, error, isCreate }) => {
   return isCreate && (
     <div className={styles.root}>
       <h3 className={styles.heading}>
-        <Avatar style={{ marginRight: 8 }} size={40} username={data?.user?.username} url={data?.user?.profilePicture} />
+        <Avatar
+          style={{ marginRight: 8 }}
+          size={40}
+          alt={data?.user?.username}
+          src={data?.user?.profilePicture}
+          icon={<UserOutlined />}
+        />
         {data?.user ? `What's on your post, ${data?.user.name}?` : 'Share your posts'}
       </h3>
       {loading ? (