+++ CHANGELOG.md
@@ -26,3 +26,6 @@
 
 ##### Release date: 15/6/2022
 - improvement: Modify UI
+
+##### Release date: 16/6/2022
+- improvement: Modify create and edit post form